smart robot trader for cryptocurrency
💎A smart robot trade that will leverage and without leverage (of your choice) and earn only 8% profit of your profit from each trade. And if your trade rarely stop limit, you will be reimbursed for loss in subsequent trades by not receiving a 8% commission.

ً🚨Warning: Our company is operating under the FATF rules and restricted countries cannot use this robot's features
Limited countries: North Korea and Iran

Our robot uses process trendlines and uses Ichimoku, RSI and Momentum as well as a proprietary template. We will always apply the overall market and Funda trends to the robot to maximize returns.

✅free(8% commission of your profit! - app free)

✅safe

✅powerful 

✅smart

Using powerful technical and Funda analytical tools has generated a 95% gain in the system, whether it is downtrend or uptrend.

You can test this robot for a month without depositing whith the unrealistic 0.1 bitcoin that the robot will give you, and if you want to invest, you can deposit up to a maximum of bitcoin(up to 100 bitcoins) or the equivalent of ethereum or tether or any coin of our systems. you can Change the bitcoin and you can select the desired currency when withdrawal.